ip_project/requirements.txt
django
django-import_export
djangorestframework